delete session flow during logout fixes

+++ src/routes/devices-router.ts
@@ -3,6 +3,7 @@ import { devicesService } from '../domain/devices-service';
 import { STATUS_CODES } from '../helpers/StatusCodes';
 import { validateRefreshToken } from '../middlewares/auth-middleware';
 import { DeviceViewModel } from '../models/DeviceModels/DeviceViewModel';
+import { RequestWithParams } from '../types/types';
 
 export const devicesRouter = Router({});
 
@@ -21,12 +22,26 @@ devicesRouter.delete('/devices', validateRefreshToken, async (req: Request, res:
   }
 });
 
-devicesRouter.delete('/devices/:id', validateRefreshToken, async (req: Request, res: Response) => {
-  const { id } = req.params;
-  const foundDevice = await devicesService.deleteDeviceById(id);
-  if (!foundDevice) {
-    res.sendStatus(STATUS_CODES.NOT_FOUND);
-    return;
-  }
-  res.sendStatus(STATUS_CODES.NO_CONTENT);
-});
+devicesRouter.delete(
+  '/devices/:id',
+  validateRefreshToken,
+  async (req: RequestWithParams<{ id: string }>, res: Response) => {
+    const { id } = req.params;
+    const currentUserId = req.user?.userId;
+    const device = await devicesService.getDeviceById(id);
+    if (!device) {
+      res.sendStatus(STATUS_CODES.NOT_FOUND);
+      return;
+    }
+    if (currentUserId !== device.userId) {
+      res.sendStatus(STATUS_CODES.FORBIDDEN);
+      return;
+    }
+    const deletedDevice = await devicesService.deleteDeviceById(id);
+    if (!deletedDevice) {
+      res.sendStatus(STATUS_CODES.NOT_FOUND);
+      return;
+    }
+    res.sendStatus(STATUS_CODES.NO_CONTENT);
+  },
+);
